Finalized Tank Search Page

+++ DOWTank.Core/Domain/TANK_usp_sel/TANK_usp_sel_DispatchReasonDDL_spResults.cs
@@ -8,7 +8,7 @@ namespace DOWTank.Core.Domain.TANK_usp_sel
 {
     public class TANK_usp_sel_DispatchReasonDDL_spResults
     {
-        public Int32 DispatchReasonTypeCD { get; set; }
+        public Int16 DispatchReasonTypeCD { get; set; }
         public string DispatchReasonTypeDS { get; set; }
     }
 }
